+++ packages/infra/src/logger/logFmtLogger.ts
@@ -1,11 +1,11 @@
 import { HashMap, Logger } from "effect-app"
 import { spanAttributes } from "../RequestContext.js"
-import { getRequestContextFromFiberRefs } from "./shared.js"
+import { getRequestContextFromCurrentContext } from "./shared.js"
 
 export const logfmtLogger = Logger.make<unknown, void>(
   (_) => {
     let { annotations } = _
-    const requestContext = getRequestContextFromFiberRefs(_.context)
+    const requestContext = getRequestContextFromCurrentContext(_.context)
     if (requestContext && requestContext.name !== "_root_") {
       annotations = HashMap.make(...[
         ...annotations,